+++ lib/karafka/web/ui/models/message.rb
@@ -7,6 +7,8 @@ module Karafka
         # A proxy between `::Karafka::Messages::Message` and web UI
         # We work with the Karafka messages but use this model to wrap the work needed.
         class Message
+          extend Lib::Paginations::Paginators
+
           class << self
             # Looks for a message from a given topic partition
             #
@@ -37,12 +39,12 @@ module Karafka
             # @param topic_id [String]
             # @param partition_id [Integer]
             # @param start_offset [Integer] oldest offset from which we want to get the data
-            # @param low_offset [Integer] low watermark offset
-            # @param high_offset [Integer] high watermark offset
+            # @param watermark_offsets [Ui::Models::WatermarkOffsets] watermark offsets
             # @return [Array] We return page data as well as all the details needed to build
             #   the pagination details.
-            def offset_page(topic_id, partition_id, start_offset, low_offset, high_offset)
-              partitions_count = fetch_partition_count(topic_id)
+            def offset_page(topic_id, partition_id, start_offset, watermark_offsets)
+              low_offset = watermark_offsets.low
+              high_offset = watermark_offsets.high
 
               # If we start from offset -1, it means we want first page with the most recent
               # results. We obtain this page by using the offset based on the high watermark
@@ -50,7 +52,7 @@ module Karafka
               start_offset = high_offset - per_page if start_offset == -1
 
               # No previous pages, no data, and no more offsets
-              no_data_result = [false, [], false, partitions_count]
+              no_data_result = [false, [], false]
 
               # If there is no data, we return the no results result
               return no_data_result if low_offset == high_offset
@@ -102,13 +104,89 @@ module Karafka
                   # If there is a potential previous page with more recent data, compute its
                   # offset
                   previous_offset >= high_offset ? false : previous_offset,
-                  fill_compacted(messages, context_offset, context_count).reverse,
-                  next_offset,
-                  partitions_count
+                  fill_compacted(messages, partition_id, context_offset, context_count).reverse,
+                  next_offset
                 ]
               end
 
               no_data_result
+            end
+
+            # Fetches requested `page_count` number of Kafka messages from the topic partitions
+            # and merges the results. Ensures, that pagination works as expected.
+            #
+            # @param topic_id [String]
+            # @param partitions_ids [Array<Integer>] for which of the partitions we want to
+            #   get the data. This is a limiting factor because of the fact that we have to
+            #   query the watermark offsets independently
+            # @param page [Integer] which page we want to get
+            def topic_page(topic_id, partitions_ids, page)
+              # This is the bottleneck, for each partition we make one request :(
+              offsets = partitions_ids.map do |partition_id|
+                [partition_id, Models::WatermarkOffsets.find(topic_id, partition_id)]
+              end.to_h
+
+              # Count number of elements we have in each partition
+              # This assumes linear presence until low. If not, gaps will be filled like we fill
+              # for per partition view
+              counts = offsets.values.map { |offset| offset[:high] - offset[:low] }
+
+              # Establish initial offsets for the iterator (where to start) per partition
+              # We do not use the negative lookup iterator because we already can compute starting
+              # offsets. This saves a lot of calls to Kafka
+              ranges = Sets.call(counts, page).map do |partition_position, partition_range|
+                partition_id = partitions_ids.to_a[partition_position]
+                watermarks = offsets[partition_id]
+
+                lowest = watermarks[:high] - partition_range.last - 1
+                # We -1 because high watermark offset is the next incoming offset and not the last
+                # one in the topic partition
+                highest = watermarks[:high] - partition_range.first - 1
+
+                # This range represents offsets we want to fetch
+                [partition_id, lowest..highest]
+              end.to_h
+
+              # We start on our topic from the lowest offset for each expected partition
+              iterator = Karafka::Pro::Iterator.new(
+                { topic_id => ranges.transform_values(&:first) }
+              )
+
+              # Build the aggregated representation for each partition messages, so we can start
+              # with assumption that all the topics are fully compacted. Then we can nicely replace
+              # compacted `false` data with real messages, effectively ensuring that the gaps are
+              # filled with `false` out-of-the-box
+              aggregated = Hash.new { |h, k| h[k] = {} }
+
+              # We initialize the hash so we have a constant ascending order based on the partition
+              # number
+              partitions_ids.each { |i| aggregated[i] }
+
+              # We prefill all the potential offsets for each partition, so in case they were
+              # compacted, we get a continuous flow
+              ranges.each do |partition, range|
+                partition_aggr = aggregated[partition]
+                range.each { |i| partition_aggr[i] = [partition, i] }
+              end
+
+              # Iterate over all partitions and collect data
+              iterator.each do |message|
+                range = ranges[message.partition]
+
+                # Do not fetch more data from a partition for which we got last message from the
+                # expected offsets
+                # When all partitions are stopped, we will stop operations. This drastically
+                # improves performance because we no longer have to poll nils
+                iterator.stop_current_partition if message.offset >= range.last
+
+                partition = aggregated[message.partition]
+                partition[message.offset] = message
+              end
+
+              [
+                aggregated.values.map(&:values).map(&:reverse).reduce(:+),
+                !Sets.call(counts, page + 1).empty?
+              ]
             end
 
             private
@@ -124,16 +202,6 @@ module Karafka
               raise
             end
 
-            # @param topic_id [String] id of the topic
-            # @return [Integer] number of partitions this topic has
-            def fetch_partition_count(topic_id)
-              ::Karafka::Admin
-                .cluster_info
-                .topics
-                .find { |topic| topic[:topic_name] == topic_id }
-                .fetch(:partition_count)
-            end
-
             # @return [Integer] elements per page
             def per_page
               ::Karafka::Web.config.ui.per_page
@@ -143,16 +211,17 @@ module Karafka
             # we need to fill those with  just the missing offset and handle this on the UI.
             #
             # @param messages [Array<Karafka::Messages::Message>] selected messages
+            # @param partition_id [Integer] number of partition for which we fill message gap
             # @param start_offset [Integer] offset of the first message (lowest) that we received
             # @param count [Integer] how many messages we wanted - we need that to fill spots to
             #   have exactly the number that was  requested and not more
             # @return [Array<Karafka::Messages::Message, Integer>] array with gaps filled with the
             #   missing offset
-            def fill_compacted(messages, start_offset, count)
+            def fill_compacted(messages, partition_id, start_offset, count)
               Array.new(count) do |index|
                 messages.find do |message|
                   (message.offset - start_offset) == index
-                end || start_offset + index
+                end || [partition_id, start_offset + index]
               end
             end
           end